+++ Glosa.Net/Core/Interfaces/IVector.cs
@@ -12,23 +12,23 @@ namespace Glosa.Net.Core.Interfaces
         void AddSelf(T vector);
         T SubtractNew(T vector);
         void SubtractSelf(T vector);
-        T DivideNew(float f);
-        void DivideSelf(float f);
-        T MultiplyNew(float f);
-        void MultiplySelf(float f);
+        T DivideNew(double f);
+        void DivideSelf(double f);
+        T MultiplyNew(double f);
+        void MultiplySelf(double f);
         //float Cross(T vector); //vector2 returns float and 3 returns vector, do we need float?
-        float Dot(T vector);
+        double Dot(T vector);
         T InverseNew();
         void InverseSelf();
-        float Heading();
+        double Heading();
         T ReflectNew(T vector);
         void ReflectSelf(T vector);
-        T RefractNew(T vector, float f);
-        void RefractSelf(T vector, float f);
-        T NormalizeNew(float v);
-        void NormalizeSelf(float v);
-        float Magnitude();
-        float AngleBetween(T vector);
-        void Set(float n);
+        T RefractNew(T vector, double f);
+        void RefractSelf(T vector, double f);
+        T NormalizeNew(double v);
+        void NormalizeSelf(double v);
+        double Magnitude();
+        double AngleBetween(T vector);
+        void Set(double n);
     }
 }
